pyproject.toml
[tool.poetry]
name = "fireswag"
version = "0.1"
description = "Auto-Setup & CSS Styling for Firefox"
authors = ["fschlegelone"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.43"
[build-system]
requires = [ "poetry-core>=1.0.0" ]
build-backend = "poetry.core.masonry.api"
